[refact] Seperate CommandBuffer functions into seperate Class  - also sperated SSDContext into seperate files
