+++ CRAProject_SSD/ssdDriver.cpp
@@ -1,8 +1,14 @@
+#pragma once
+
 #include <vector>
 #include <string>
 #include <fstream>
 #include <sstream>
 #include <filesystem>
+
+
+#include "ssdContext.cpp"
+#include "commandBuffer.cpp"
 
 using namespace std;
 using namespace std::filesystem;
@@ -13,38 +19,6 @@ public:
     virtual void execute() = 0;
 
     const int LBA_MAX = 100;
-};
-
-struct SSDContext {
-    fstream nand;
-    const string nandFileName = "ssd_nand.txt";
-
-    static void overwriteTextToFile(const string& fileName, const string& text) {
-        ofstream file(fileName);
-        if (!file.is_open()) return;
-        
-        file << text;
-        file.close();
-    }
-
-    bool openOrCreateNand(ios_base::openmode mode) {
-        nand.open(nandFileName, mode);
-        if (!nand.is_open()) {
-            ofstream createFile(nandFileName);
-            createFile.close();
-            nand.open(nandFileName, mode);
-        }
-        return nand.is_open();
-    }
-
-    string handleErrorReturn() {
-        overwriteTextToFile("ssd_output.txt", "ERROR");
-        return "";
-    }
-
-    void handleError() {
-        overwriteTextToFile("ssd_output.txt", "ERROR");
-    }
 };
 
 class WriteCommand : public Command {
@@ -156,8 +130,8 @@ public:
 
         if (argc <= 1) return ctx.handleError();
 
-        vector<string> fileNames = createAndReadFiles();
-        vector<vector<string>> buffer = parseFileNamse(fileNames);
+        vector<string> fileNames = commandBufferManager.createAndReadFiles();
+        vector<vector<string>> buffer = commandBufferManager.parseFileNames(fileNames);
 
         vector<string> args = parseArguments(argc, argv);
         string command = args[0];
@@ -320,7 +294,7 @@ public:
                 return ctx.handleError();
             }
             
-            writeCommandBuffer(buffer);
+            commandBufferManager.writeCommandBuffer(buffer);
         }
         catch (...) {
             return ctx.handleError();
@@ -331,6 +305,7 @@ public:
 
 private:
     SSDContext ctx;
+    CommandBufferManager commandBufferManager{ctx, "./buffer" };
 
     static vector<string> parseArguments(int argc, char* argv[]) {
         vector<string> args;
@@ -338,123 +313,6 @@ private:
             args.emplace_back(argv[i]);
         }
         return args;
-    }
-
-    vector<string> createAndReadFiles(void)
-    {
-        vector<string> names;
-        try {
-            string dirPath = "./buffer";
-    
-            // Check And Create
-            path p(dirPath);
-            if (!exists(p)) {
-                if (!create_directory(p)) {
-                    throw runtime_error("Failed to create directory: " + dirPath);
-                }
-            }
-
-            vector<string> checkFiles;
-            for (const auto& entry : directory_iterator(dirPath)) {
-                if (entry.is_directory()) {
-                    checkFiles.push_back(entry.path().filename().string());
-                }
-            }
-
-            if (checkFiles.size() == 0) {
-                for (int i = 1; i <= 5; ++i) {
-                    string fileName = to_string(i) + "_empty";
-                    path filePath = p / fileName;
-
-                    ofstream ofs(filePath);
-                    if (!ofs) {
-                        throw runtime_error("Failed to create file: " + filePath.string());
-                    }
-                }
-            }
-
-            // Read
-            for (const auto& entry : directory_iterator(dirPath)) {
-                if (!entry.is_regular_file()) continue;
-
-                names.push_back(entry.path().filename().string());
-            }
-        }
-        catch (const exception& err) {
-            ctx.handleError();
-            return vector<string>();
-        }
-
-        return names;
-    }
-
-    vector<string> split(const string& str, char delimiter = '_') {
-        vector<string> tokens;
-        string token;
-        stringstream ss(str);
-
-        while (getline(ss, token, delimiter)) {
-            tokens.push_back(token);
-        }
-
-        return tokens;
-    }
-
-    vector<vector<string>> parseFileNamse(vector<string> fileNames)
-    {
-        vector<vector<string>> result;
-        for (const string& fileName : fileNames)
-        {
-            vector<string> fileNameSpliteed = split(fileName);
-            fileNameSpliteed.erase(fileNameSpliteed.begin());
-
-            if (fileNameSpliteed[0] == "empty") continue;
-
-            result.push_back(fileNameSpliteed);
-        }
-
-        return result;
-    }
-
-    void addCommandBuffer(vector<vector<string>>& buffer, vector<string> args)
-    {
-        buffer.push_back(args);
-    }
-
-    string joinStrings(const vector<string> vec, const string delimiter = "_") {
-        string result;
-        for (size_t i = 0; i < vec.size(); ++i) {
-            result += vec[i];
-            if (i != vec.size() - 1) {
-                result += delimiter;
-            }
-        }
-        return result;
-    }
-
-    void writeCommandBuffer(vector<vector<string>> buffer)
-    {
-        while (buffer.size() < 5) buffer.push_back(vector<string>({ "empty" }));
-
-        path dirPath = "./buffer";
-        for (const auto& entry : directory_iterator(dirPath)) {
-            if (entry.is_regular_file()) {
-                std::error_code ec;
-                remove(entry.path(), ec);
-                if (ec) return ctx.handleError();
-            }
-        }
-
-        for (int i = 0; i < buffer.size(); i++) {
-            string fileName = to_string(i + 1) + "_" +  joinStrings(buffer[i]);
-
-            path filePath = dirPath / fileName;
-
-            ofstream ofs(filePath);
-            if (!ofs) return ctx.handleError();
-            
-            ofs.close();
-        }
     }
 
     bool fastRead(vector<string> args, vector<vector<string>> buffer)
